ignore: Test case cleanup
Some serialization tests was in several files.
Using snake-case for folder names for resource test files
