+++ v16/src/test/kotlin/com/monta/library/ocpp/v16/core/RemoteStopTransactionSerializationTest.kt
@@ -14,7 +14,7 @@ class RemoteStopTransactionSerializationTest : StringSpec({
     val messageSerializer = MessageSerializer(SerializationMode.OCPP_1_6, OcppErrorResponderV16)
 
     "parse RemoteStopTransaction request" {
-        val jsonString = TestUtils.getFileAsString("remotestoptransaction/req.json")
+        val jsonString = TestUtils.getFileAsString("remote_stop_transaction/req.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Request>>()
         val ocppMessage = parsingResult.value
@@ -34,7 +34,7 @@ class RemoteStopTransactionSerializationTest : StringSpec({
     }
 
     "parse RemoteStopTransaction response" {
-        val jsonString = TestUtils.getFileAsString("remotestoptransaction/res.json")
+        val jsonString = TestUtils.getFileAsString("remote_stop_transaction/res.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Response>>()
         val ocppMessage = parsingResult.value
