+++ v16/src/test/kotlin/com/monta/library/ocpp/v16/core/StopTransactionSerializationTest.kt
@@ -10,6 +10,9 @@ import com.monta.library.ocpp.v16.IdTagInfo
 import com.monta.library.ocpp.v16.SampledValue
 import com.monta.library.ocpp.v16.error.OcppErrorResponderV16
 import io.kotest.core.spec.style.StringSpec
+import io.kotest.matchers.collections.shouldBeEmpty
+import io.kotest.matchers.nulls.shouldBeNull
+import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import java.time.ZonedDateTime
@@ -17,8 +20,30 @@ import java.time.ZonedDateTime
 class StopTransactionSerializationTest : StringSpec({
     val messageSerializer = MessageSerializer(SerializationMode.OCPP_1_6, OcppErrorResponderV16)
 
+    "parse a siemens stoptransaction with partly invalid transaction data (CPIBUG-61)" {
+        val jsonString = TestUtils.getFileAsString("stop_transaction/stoptransaction-siemens.json")
+        val parsingResult = messageSerializer.parse(jsonString)
+
+        parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Request>>()
+        val ocppMessage = parsingResult.value
+        ocppMessage.shouldBeInstanceOf<Message.Request>()
+        ocppMessage.uniqueId shouldBe "1700035413777"
+        ocppMessage.action shouldBe "StopTransaction"
+
+        val stopTransactionParsingResult = messageSerializer.deserializePayload(ocppMessage, StopTransactionRequest::class.java)
+        stopTransactionParsingResult.shouldBeInstanceOf<ParsingResult.Success<StopTransactionRequest>>()
+        val stopTransaction = stopTransactionParsingResult.value
+
+        val transactionData = stopTransaction.transactionData.shouldNotBeNull()
+        transactionData.size shouldBe 2
+        transactionData.forEach {
+            it.timestamp.shouldBeNull()
+            it.sampledValue.shouldBeEmpty()
+        }
+    }
+
     "parse StopTransaction request with only mandatory fields" {
-        val jsonString = TestUtils.getFileAsString("stoptransaction/req.json")
+        val jsonString = TestUtils.getFileAsString("stop_transaction/req.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Request>>()
         val ocppMessage = parsingResult.value
@@ -49,7 +74,7 @@ class StopTransactionSerializationTest : StringSpec({
     }
 
     "parse StopTransaction request with all optional fields" {
-        val jsonString = TestUtils.getFileAsString("stoptransaction/req_optional.json")
+        val jsonString = TestUtils.getFileAsString("stop_transaction/req_optional.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Request>>()
         val ocppMessage = parsingResult.value
@@ -90,7 +115,7 @@ class StopTransactionSerializationTest : StringSpec({
     }
 
     "parse StopTransaction response with only mandatory fields" {
-        val jsonString = TestUtils.getFileAsString("stoptransaction/res.json")
+        val jsonString = TestUtils.getFileAsString("stop_transaction/res.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Response>>()
         val ocppMessage = parsingResult.value
@@ -109,7 +134,7 @@ class StopTransactionSerializationTest : StringSpec({
     }
 
     "parse StopTransaction response with all optional fields" {
-        val jsonString = TestUtils.getFileAsString("stoptransaction/res_optional.json")
+        val jsonString = TestUtils.getFileAsString("stop_transaction/res_optional.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Response>>()
         val ocppMessage = parsingResult.value
