+++ v16/src/test/kotlin/com/monta/library/ocpp/v16/core/UnlockConnectorSerializationTest.kt
@@ -14,7 +14,7 @@ class UnlockConnectorSerializationTest : StringSpec({
     val messageSerializer = MessageSerializer(SerializationMode.OCPP_1_6, OcppErrorResponderV16)
 
     "parse UnlockConnector request" {
-        val jsonString = TestUtils.getFileAsString("unlockconnector/req.json")
+        val jsonString = TestUtils.getFileAsString("unlock_connector/req.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Request>>()
         val ocppMessage = parsingResult.value
@@ -34,7 +34,7 @@ class UnlockConnectorSerializationTest : StringSpec({
     }
 
     "parse UnlockConnector response" {
-        val jsonString = TestUtils.getFileAsString("unlockconnector/res.json")
+        val jsonString = TestUtils.getFileAsString("unlock_connector/res.json")
         val parsingResult = messageSerializer.parse(jsonString)
         parsingResult.shouldBeInstanceOf<ParsingResult.Success<Message.Response>>()
         val ocppMessage = parsingResult.value
