+++ v16/src/test/resources/boot_notification/req_optional.json
