+++ v16/src/test/resources/boot_notification/res.json
