+++ v16/src/test/resources/change_availability/req.json
