+++ v16/src/test/resources/clear_cache/req.json
