+++ v16/src/test/resources/clear_cache/res.json
