+++ v16/src/test/resources/data_transfer/res_optional.json
