+++ v16/src/test/resources/meter_values/req_optional.json
