+++ v16/src/test/resources/meter_values/res.json
