+++ v16/src/test/resources/remote_stop_transaction/res.json
