--- v16/src/test/resources/status_notification/err.json
@@ -1,7 +1,0 @@
-[
-  4,
-  "78ee4811-a044-417e-9c70-b51c2c296bf9",
-  "This is a code",
-  "This is a description",
-  {}
-]
