+++ v16/src/test/resources/status_notification/req_optional.json
