+++ v16/src/test/resources/status_notification/res.json
@@ -1,5 +1,1 @@
-[
-  3,
-  "78ee4811-a044-417e-9c70-b51c2c296bf9",
-  {}
-]
+[3,"res-1",{}]
