--- v16/src/test/resources/statusnotification/req.json
@@ -1,1 +1,0 @@
-[2,"req-1","StatusNotification",{"connectorId":1,"errorCode":"NoError","status":"Available"}]
