+++ v16/src/test/resources/stop_transaction/req.json
