+++ v16/src/test/resources/stop_transaction/stoptransaction-siemens.json
