--- v16/src/test/resources/stoptransaction/res.json
@@ -1,1 +1,0 @@
-[3,"res-1",{}]
